Conditionally apply the latest tag
This commit adds a condition to apply the latest
tag only if the branch is master. This is done to
avoid applying the latest tag on other branches.
Also the latest tag is the 2nd tag so that we can
consume the first one for scanning the image with
trivy.

Signed-off-by: Suraj Deshmukh <surajd.service@gmail.com>